ml-service/requirements-gcp.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers>=2.3.0
huggingface-hub>=0.19.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
kagglehub==0.2.0
pymongo==4.6.0
python-multipart==0.0.6
google-cloud-storage>=2.10.0




